# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API aplikasi flashcard untuk mengingat nama mahasiswa seangkatan: kelola mahasiswa, main flashcard (skor +2/-1), dan kelola self-reward (klaim → reset poin).",
        "title": "FlashCard Angkatan API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/play/answer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play"
                ],
                "summary": "Jawab pertanyaan",
                "parameters": [
                    {
                        "description": "Jawaban",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "correct, points_delta, current_points, correct_name",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Kartu tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/end": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play"
                ],
                "summary": "Akhiri sesi \u0026 catat statistik",
                "parameters": [
                    {
                        "description": "Hasil sesi",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "session_id, points_earned, accuracy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid / tidak konsisten",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/round": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play"
                ],
                "summary": "Ambil kartu acak (round baru)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kartu yang sudah dimainkan, pisahkan koma (1,2,3)",
                        "name": "exclude",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "card_id, photo_url, options[4] — {finished:true} jika sesi habis",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Minimal 4 mahasiswa berfoto",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Kesalahan server",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/rewards/{id}/claim": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rank"
                ],
                "summary": "Klaim reward",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID reward",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "claimed_reward + poin direset ke 0",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Poin belum mencukupi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Reward tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    }
}